+++ Projeto_de_Programa-o_N04_Grupo-01/backend/src/main/java/br/com/unit/controller/StaffController.java
@@ -31,7 +31,7 @@ public class StaffController {
     }
 
     @Deletemapping("/staff/deletar/{id}")
-    public String removerStaff(@PathVariable int id){
+    public String removerStaff(@PathVariable long id){
         staff.removeIf(e -> e.getIdStaff() == id);
         return "Membro do staff com ID " + id + " removido com sucesso!";
     }
